Update README.md to include --device cpu
Based on this: https://github.com/m-bain/whisperX/issues/748

I am not sure if on OS X, only CPU is possible or if this should be split into too (ie. how to run on OS X and how to run on CPU). Without `--device cpu`, I was getting the libcudnn8 error, so it was still trying to run on GPU.